+++ src/main/kotlin/me/nobaboy/nobaaddons/commands/debug/ItemDebugCommands.kt
@@ -1,25 +1,39 @@
 package me.nobaboy.nobaaddons.commands.debug
 
+import dev.celestialfault.commander.annotations.Command
+import dev.celestialfault.commander.annotations.Group
+import dev.celestialfault.commander.annotations.RootCommand
 import me.nobaboy.nobaaddons.commands.debug.DebugCommands.dumpInfo
-import me.nobaboy.nobaaddons.commands.internal.Command
-import me.nobaboy.nobaaddons.commands.internal.Group
+import me.nobaboy.nobaaddons.commands.impl.Context
 import me.nobaboy.nobaaddons.utils.MCUtils
+import me.nobaboy.nobaaddons.utils.TextUtils.buildText
+import me.nobaboy.nobaaddons.utils.TextUtils.darkAqua
+import me.nobaboy.nobaaddons.utils.TextUtils.toText
 import me.nobaboy.nobaaddons.utils.items.ItemUtils.asSkyBlockItem
 import me.nobaboy.nobaaddons.utils.items.ItemUtils.isSkyBlockItem
+import net.minecraft.command.argument.NbtPathArgumentType
 import net.minecraft.component.DataComponentTypes
+import net.minecraft.component.type.LoreComponent
+import net.minecraft.component.type.NbtComponent
+import net.minecraft.nbt.NbtElement
+import net.minecraft.nbt.NbtHelper
+import net.minecraft.nbt.NbtList
+import net.minecraft.nbt.NbtString
 import net.minecraft.text.Text
 
 @Suppress("unused")
-object ItemDebugCommands : Group("item") {
-	override val root = RootCommand {
+@Group("item")
+object ItemDebugCommands {
+	@RootCommand
+	fun root(ctx: Context) {
 		val item = MCUtils.player!!.mainHandStack
 		if(item.isEmpty || !item.isSkyBlockItem) {
-			it.source.sendError(Text.literal("You aren't holding a valid SkyBlock item"))
-			return@RootCommand
+			ctx.source.sendError(Text.literal("You aren't holding a valid SkyBlock item"))
+			return
 		}
 
 		val itemData = item.asSkyBlockItem!!
-		it.dumpInfo(
+		ctx.dumpInfo(
 			"Item ID" to itemData.id,
 			"UUID" to itemData.uuid,
 			"Created" to itemData.timestamp?.elapsedSince(),
@@ -38,25 +52,46 @@ object ItemDebugCommands : Group("item") {
 		)
 	}
 
-	val dumpNbt = Command("nbt") {
+	@Command
+	fun nbt(ctx: Context, path: NbtPathArgumentType.NbtPath? = null) {
 		val item = MCUtils.player!!.mainHandStack
 		if(item.isEmpty) {
-			it.source.sendError(Text.literal("You aren't holding an item"))
-			return@Command
+			ctx.source.sendError(Text.literal("You aren't holding an item"))
+			return
 		}
 
-		println(item.get(DataComponentTypes.CUSTOM_DATA))
-		it.source.sendFeedback(Text.literal("Dumped item NBT to game logs"))
+		var nbt: NbtElement = item.getOrDefault(DataComponentTypes.CUSTOM_DATA, NbtComponent.DEFAULT).copyNbt()
+		if(path != null) {
+			nbt = path.get(nbt).first()
+		}
+
+		ctx.source.sendFeedback(buildText {
+			append("Item NBT: ".toText().darkAqua())
+			append(NbtHelper.toPrettyPrintedText(nbt))
+		})
 	}
 
-	val dumpLore = Command("lore") {
+	@Command
+	fun lore(ctx: Context, line: Int? = null) {
 		val item = MCUtils.player!!.mainHandStack
 		if(item.isEmpty || !item.contains(DataComponentTypes.LORE)) {
-			it.source.sendError(Text.literal("You aren't holding an item with lore"))
-			return@Command
+			ctx.source.sendError(Text.literal("You aren't holding an item with lore"))
+			return
 		}
 
-		println(item.get(DataComponentTypes.LORE)!!.lines)
-		it.source.sendFeedback(Text.literal("Dumped item lore to game logs"))
+		var nbt: NbtElement = NbtList().apply {
+			item.getOrDefault(DataComponentTypes.LORE, LoreComponent.DEFAULT).lines.forEach {
+				add(NbtString.of(it.string))
+			}
+		}
+		if(line != null) {
+			val path = NbtPathArgumentType.NbtPath.parse("[$line]")
+			nbt = path.get(nbt).first()
+		}
+
+		ctx.source.sendFeedback(buildText {
+			append("Item lore: ".toText().darkAqua())
+			append(NbtHelper.toPrettyPrintedText(nbt))
+		})
 	}
 }